requirements.txt
flask==2.0.1
openai==0.27.0
python-dotenv==0.19.0
requests==2.28.1
argparse==1.4.0